lan-file-manager/backend/requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
pyftpdlib==1.5.9
python-multipart==0.0.6
python-jose[cryptography]==3.3.0
passlib==1.7.4
bcrypt==3.2.2
aiofiles==23.2.1
websockets==12.0
pydantic==2.5.3
pydantic-settings==2.1.0
